评论模块重构。

+++ src/main/java/com/suanfa8/algocrazyapi/dto/comment/CommentDeleteDto.java
@@ -11,8 +11,8 @@ import lombok.ToString;
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class CommentDeleteDto {
 
-    private Integer commentId;
+    private Long commentId;
 
-    private Integer parentCommentId;
+    private Long parentCommentId;
 
 }
