+++ src/main/java/com/suanfa8/algocrazyapi/mapper/CommentMapper.java
@@ -12,7 +12,7 @@ public interface CommentMapper extends BaseMapper<Comment> {
 
     // List<Comment> getCommentsByArticleId(Integer articleId);
 
-    default boolean incrementLikeCount(Integer id) {
+    default boolean incrementLikeCount(Long id) {
         LambdaUpdateWrapper<Comment> updateWrapper = new LambdaUpdateWrapper<>();
         updateWrapper.eq(Comment::getId, id).setSql("like_count = like_count + 1");
         return update(updateWrapper) > 0;
